+++ packages/rax-image/src/index.tsx
@@ -1,6 +1,6 @@
 import { isWeb, isWeex, isMiniApp, isWeChatMiniProgram, isByteDanceMicroApp, isBaiduSmartProgram, isKuaiShouMiniProgram } from 'universal-env';
 import ImageWeb from './web';
-import ImageMiniApp from './miniapp';
+import ImageMiniApp from './miniapp-runtime';
 import ImageWeex from './weex';
 
 let Image = null;