+++ packages/rax-modal/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.6.3
+
+- Keep miniapp native code in older dir
+
 ## 1.6.2
 
 - Reduce npm files